lab/pi/ssd1306_dual/requirements.txt
# Install on the Raspberry Pi (venv or --user). Dev-Host does not need these for repo tests.
luma.oled>=3.13,<4
pillow>=10,<12
